Refactored how config file is handled
While attempting to add validation on parameters like client ID, secret, etc., I realized that because those values could be either in a parameter OR in the config file, you couldn't just throw an error if a parameter was missing.

To solve this, I created a custom binder. This allows our command handler to just take an instance of the CredentialsOptions class that is built from the appropriate combo of command line parameters and values in the config file. The binder takes care of which value takes precedence.

I added a ValidateOptionsForFlow method to the CredentialOptions class that throws an exception if the options aren't valid, based on which auth flow is requested. For example, a client secret is required for client credentials, but not for others.